Extract hooks to separate modules
