+++ assets/js/VolumeSlider.js
@@ -1,0 +1,7 @@
+export default {
+  mounted() {
+    this.el.addEventListener("change", (e) => {
+      this.pushEvent("set_volume", { volume_percent: e.target.valueAsNumber });
+    });
+  },
+};
